fix: 🐛 change number fields in WeaponPerformanceStats to long to fix broken query in Splatoon3VsWeaponRepository

+++ src/main/java/tv/strohi/twitch/strohkoenigbot/splatoon3saver/database/repo/vs/model/WeaponPerformanceStats.java
@@ -10,13 +10,13 @@ import javax.persistence.Cacheable;
 public class WeaponPerformanceStats {
 	private final Splatoon3VsWeapon weapon;
 	private final boolean isMyTeam;
-	private final int totalGames;
-	private final int totalWins;
-	private final int totalDefeats;
+	private final long totalGames;
+	private final long totalWins;
+	private final long totalDefeats;
 	private final double winRate;
 	private final double defeatRate;
 
-	public WeaponPerformanceStats(Splatoon3VsWeapon weapon, boolean isMyTeam, int totalGames, int totalWins, int totalDefeats) {
+	public WeaponPerformanceStats(Splatoon3VsWeapon weapon, boolean isMyTeam, long totalGames, long totalWins, long totalDefeats) {
 		this.weapon = weapon;
 		this.isMyTeam = isMyTeam;
 		this.totalGames = totalGames;
